moved code to do aggregations higher up inheritance tree
